+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ## 1.1.0
 
 - Support Flutter SDK
+- Dart and Flutter example